6. Подсчет функции в диапазоне. [#24251 #388059]

+++ src/main/java/ru/job4j/function/FunctionCalculator.java
@@ -5,10 +5,10 @@ import java.util.List;
 import java.util.function.Function;
 
 public class FunctionCalculator {
-    public List<Double> diapason(int start, int end, Function<Integer, Double> func) {
+    public List<Double> diapason(int start, int end, Function<Double, Double> func) {
         List<Double> rsl = new ArrayList<>();
         for (int i = start; i < end; i++) {
-            rsl.add(func.apply(i));
+            rsl.add(func.apply((double) i));
         }
         return rsl;
     }
